requirements.txt
SQLAlchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
PySide6>=6.7
alembic>=1.13
reportlab>=4.0
pillow>=10.0
pytest>=8.0
pyinstaller>=6.0